chess-nn/requirements.txt
torch>=2.0.0
numpy>=1.24.0
python-chess>=1.9.0
requests>=2.28.0
zstandard>=0.21.0
onnx>=1.14.0
onnxruntime>=1.15.0
